requirements.txt
# Core bot deps
python-telegram-bot
yfinance
openai
python-dotenv
ib_insync
pysqlite3-binary
textblob
apscheduler